requirements.txt
absl-py==2.1.0
addict==2.4.0
cachetools==5.5.0
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.3.1
cycler==0.12.1
descartes==1.1.0
einops==0.8.0
exceptiongroup==1.2.2
filelock==3.13.1
fire==0.7.0
fonttools==4.55.3
fsspec==2024.2.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
grpcio==1.70.0
h5py==3.12.1
idna==3.4
imageio==2.37.0
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.4
mmsegmentation==1.2.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.3
numpy==1.23.5
nuscenes-devkit==1.1.11
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip==24.2
pipdeptree==2.24.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.0.0
prettytable==3.12.0
protobuf==5.29.3
pycocotools==2.0.8
Pygments==2.18.0
pyparsing==3.2.0
pyquaternion==0.9.9
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
rich==13.9.4
scikit-image==0.25.1
scikit-learn==1.6.0
scipy==1.14.1
selective_scan==0.0.2
setuptools==75.1.0
Shapely==1.8.5.post1
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==0.4.12
tomli==2.2.1
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchvision==0.16.1+cu121
tqdm==4.67.1
triton==2.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.13
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.44.0
yacs==0.1.8
yapf==0.43.0
